Add files via upload
This folder includes the code used for the iEEG experiment (including triggers). I have also created the code for two control experiments: Control_exp which repeats the same DV in each repetition but using a different stimuli sequence and the Bloqued_exp, that combines blocks where the stimuli are repeated 3 times and blocks were the stimuli are randomly drawn (same DV in each repetition, but the sign is flipped randomly).