Add files via upload
This is the version of the experiment compatible with EEG and pupil labs. I have included compatibility with CEDRUX ttl port.
Participants can report the confidence in their responses.
Gratings are larger and are optimized for the eeg lab monitor.

Compatibility issues:
This experiment runs well in Psychopy21.2.3 (Python 3.8). 
In newer versions the slider and circle components look wrong.
Alexis Perez Bellido